% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_distinct_ids.R
\name{check_distinct_ids}
\alias{check_distinct_ids}
\title{Internal function which creates cross-tables with number of distinct id's}
\usage{
check_distinct_ids(data, id_var, val_var, by = NULL)
}
\arguments{
\item{data}{\link{data.frame} from which the descriptive statistics are
calculated.}

\item{id_var}{\link{character} The name of the id variable.}

\item{val_var}{\link{character} vector of value variables on which descriptive
statistics are computed.}

\item{by}{Grouping variables (or expression) as in \link{data.table}'s \code{by}.}
}
\description{
Internal function which creates cross-tables with number of distinct id's
}
