% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_log.R
\name{generate_log}
\alias{generate_log}
\title{Source a single R script and generate a log file}
\usage{
generate_log(r_script, log_file)
}
\arguments{
\item{r_script}{R script to be run and logged}

\item{log_file}{destination file of the log file to be generated}
}
\description{
Source a single R script and generate a log file
}
