% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.sdc_distinct_ids}
\alias{print.sdc_distinct_ids}
\alias{print.sdc_dominance}
\alias{print.sdc_options}
\alias{print.sdc_settings}
\alias{print.sdc_descriptives}
\alias{print.sdc_model}
\alias{print.sdc_min_max}
\title{Print methods for SDC objects}
\usage{
\method{print}{sdc_distinct_ids}(x, ...)

\method{print}{sdc_dominance}(x, ...)

\method{print}{sdc_options}(x, ...)

\method{print}{sdc_settings}(x, ...)

\method{print}{sdc_descriptives}(x, ...)

\method{print}{sdc_model}(x, ...)

\method{print}{sdc_min_max}(x, ...)
}
\arguments{
\item{x}{The object to be printed}

\item{...}{Ignored.}
}
\description{
These methods print SDC objects. Tables containing information
are only printed when relevant.
}
