% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_descriptives.R
\name{sdc_count}
\alias{sdc_count}
\title{Internal function which creates cross-tables with number of distinct id's}
\usage{
sdc_count(DT, id_var, val_var, by = NULL)
}
\arguments{
\item{DT}{The data.table from which the descriptives are calculated.}

\item{id_var}{The name of the id variable. Can be provided as bare (unquoted)
column name or as a character.}

\item{val_var}{The name of the value variable. Can be provided as bare
(unquoted) column name or as a character.}

\item{by}{Grouping variables. Can be provided as in
\code{\link[data.table:data.table]{data.table::data.table()}}.}
}
\description{
Internal function which creates cross-tables with number of distinct id's
}
