% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdc_DT}
\alias{sdc_DT}
\title{Example Datasets used in the vignette
Recreated after the example data from nobsdes.}
\format{
A data.table with 100 rows and 8 columns.
}
\usage{
data("sdc_DT")
}
\description{
Example Datasets used in the vignette
Recreated after the example data from nobsdes.
}
\details{
The data.table contains the following columns:
\itemize{
\item id \link{integer} random identifier
\item time \link{integer} random time variable
\item V1 - V3 \link{numeric} random variables
\item D1 - D3 \link{logical} non-random dummy variables
}
}
\keyword{datasets}
