% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdc_dups_credits_DT}
\alias{sdc_dups_credits_DT}
\title{Example data for \code{sdc_descriptives()}}
\format{
A data.table with 6 rows and 3 columns.
}
\usage{
data("sdc_dups_credits_DT")
}
\description{
Utilized in the unit tests
}
\details{
The data.table contains the following columns:
\itemize{
\item lender \link{factor} random lender identifier
\item borrower \link{factor} random borrower identifier
\item volume \link{integer} credit volume
}
}
\keyword{datasets}
