% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdc_dups_DT}
\alias{sdc_dups_DT}
\title{Example data for \code{sdc_descriptives()}}
\format{
A data.table with 9 rows and 5 columns.
}
\usage{
data("sdc_dups_DT")
}
\description{
Utilized in the unit tests
}
\details{
The data.table contains the following columns:
\itemize{
\item lender_group \link{factor} random lender group identifier
\item lender \link{factor} random lender identifier
\item volume \link{integer} credit volume
\item borrower \link{factor} random borrower identifier
\item borrower_group \link{factor} random borrower group identifier
}
}
\keyword{datasets}
