% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdc_dups_lender_groups_DT}
\alias{sdc_dups_lender_groups_DT}
\title{Example data for \code{sdc_descriptives()}}
\format{
A data.table with 7 rows and 2 columns.
}
\usage{
data("sdc_dups_lender_groups_DT")
}
\description{
Utilized in the unit tests
}
\details{
The data.table contains the following columns:
\itemize{
\item lender_group \link{factor} random lender group identifier
\item lender \link{factor} random lender identifier
}
}
\keyword{datasets}
