% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_DT.R
\docType{data}
\name{sdc_extreme_DT}
\alias{sdc_extreme_DT}
\title{Test Dataset for vignette}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 6 columns.
}
\usage{
sdc_extreme_DT
}
\description{
Test Dataset for vignette
}
\keyword{datasets}
