% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_log.R
\name{sdc_log}
\alias{sdc_log}
\title{Create Stata-like log files from R Scripts}
\usage{
sdc_log(r_scripts, log_files, replace = FALSE, silent = TRUE)
}
\arguments{
\item{r_scripts}{character vector containing the path(s) of the R script(s)
which should be run with logging}

\item{log_files}{character vector containing the path(s) of the text file(s)
where the log(s) should be stored}

\item{replace}{logical. Indicates whether to replace existing log files.}

\item{silent}{logical. Switch to TRUE in order to receive warnings and
messages from running the 'r_scripts'.}
}
\value{
Invisible \code{NULL}.
}
\description{
This function creates Stata-like log files from R Scripts. It
can handle single files or a list of files (in a character vector) at once.
}
