% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdc_min_max_DT}
\alias{sdc_min_max_DT}
\title{Example data for \code{sdc_min_max()}}
\format{
A data.table with 20 rows and 6 columns.
}
\usage{
data("sdc_min_max_DT")
}
\description{
Utilized in the vignette
}
\details{
The data.table contains the following columns:
\itemize{
\item id \link{factor} random identifier
\item sector \link{factor} economic sector
\item year \link{integer} time variable
\item val_1 - val_3 \link{numeric} value variables
}
}
\keyword{datasets}
