% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_DT.R
\docType{data}
\name{sdc_model_DT}
\alias{sdc_model_DT}
\title{Test Dataset for vignette}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 80 rows and 9 columns.
}
\usage{
data("sdc_DT")
}
\description{
Test Dataset for vignette
}
\keyword{datasets}
