% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warn_distinct_ids.R
\name{warn_distinct_ids}
\alias{warn_distinct_ids}
\title{Throw only a single warning about insufficient distinct ID's}
\usage{
warn_distinct_ids(list)
}
\arguments{
\item{list}{\link{list} of elements of class \code{sdc_distinct_ids}.}
}
\description{
Checks if any check on distinct ID's was problematic and throws
a single warning.
}
