% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{common_arguments}
\alias{common_arguments}
\title{arguments}
\arguments{
\item{data}{\link{data.frame} from which the descriptive statistics are
calculated.}

\item{id_var}{\link{character} The name of the id variable.}

\item{val_var}{\link{character} vector of value variables on which descriptive
statistics are computed.}

\item{by}{Grouping variables (or expression) as in \link{data.table}'s \code{by}.}

\item{zero_as_NA}{\link{logical} If TRUE, zeros in 'val_var' are treated as NA.}

\item{model}{The estimated model object. Can be a model type like lm, glm and
various others (anything which can be handled by \code{\link[broom:reexports]{broom::augment()}}).}

\item{n_min}{\link{integer} The number of values used to calculate the minimum, by
default 5.}

\item{n_max}{\link{integer} The number of values used to calculate the maximum, by
default 5.}
}
\description{
arguments
}
