% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_descriptives.R
\name{sdc_descriptives}
\alias{sdc_descriptives}
\title{Check if your descriptive statistics comply to statistical disclosure control}
\usage{
sdc_descriptives(data, id_var, val_var, by = NULL, zero_as_NA = NULL)
}
\arguments{
\item{data}{\link{data.frame} from which the descriptives are calculated.}

\item{id_var}{\link{character} The name of the id variable.}

\item{val_var}{\link{character} vector of value variables on which descriptives
are computed.}

\item{by}{Grouping variables (or expression). Can be provided as in
\code{\link[data.table:data.table]{data.table::data.table()}}.}

\item{zero_as_NA}{\link{logical} If TRUE, zeros in 'val_var' are treated as NA.}
}
\value{
A list \link{list} with detailed information about options, settings,
compliance with the criteria distinct_ids and dominance.
}
\description{
Check if your descriptive statistics comply to statistical disclosure control
}
\examples{
sdc_descriptives(
    data = sdc_descriptives_DT,
    id_var = "id",
    val_var = "val_1"
)

sdc_descriptives(
    data = sdc_descriptives_DT,
    id_var = "id",
    val_var = "val_1",
    by = sector
)

sdc_descriptives(
    data = sdc_descriptives_DT,
    id_var = "id",
    val_var = "val_1",
    by = c("sector", "year")
)

sdc_descriptives(
    data = sdc_descriptives_DT,
    id_var = "id",
    val_var = "val_2",
    by = c("sector", "year")
)

sdc_descriptives(
    data = sdc_descriptives_DT,
    id_var = "id",
    val_var = "val_2",
    by = c("sector", "year"),
    zero_as_NA = FALSE
)

}
