% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_descriptives.R
\name{sdc_descriptives}
\alias{sdc_descriptives}
\title{Check if your descriptive statistics comply to RDSC rules}
\usage{
sdc_descriptives(data, id_var, val_var, by = NULL)
}
\arguments{
\item{data}{The dataset (anything which can be coerced to data.table) from
which the descriptives are calculated.}

\item{id_var}{The name of the id variable as a character.}

\item{val_var}{Character vector of value variables on which descriptives are
computed.}

\item{by}{Grouping variables. Can be provided as in
\code{\link[data.table:data.table]{data.table::data.table()}}.}
}
\description{
Check if your descriptive statistics comply to RDSC rules
}
