% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_DT.R
\docType{data}
\name{sdc_descriptives_DT}
\alias{sdc_descriptives_DT}
\title{Test Dataset for vignette}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 5 columns.
}
\usage{
sdc_descriptives_DT
}
\description{
Test Dataset for vignette
}
\keyword{datasets}
