% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdc_descriptives_DT}
\alias{sdc_descriptives_DT}
\alias{sdc_model_DT}
\title{Example data
Utilitzed in the vignette}
\format{
A data.table with 20 rows and 4 columns.

A data.table with 80 rows and 9 columns.
}
\usage{
data("sdc_descriptives_DT")

data("sdc_model_DT")
}
\description{
Example data
Utilitzed in the vignette

Example data
Utilitzed in the vignette
}
\details{
The data.table contains the following columns:
\itemize{
\item id \link{factor} random identifier
\item sector \link{factor} economic sector
\item year \link{integer} time variable
\item val_1, val_2 \link{numeric} value variables
}

The data.table contains the following columns:
\itemize{
\item id \link{factor} random identifier
\item y - x_4 \link{numeric} value variables
\item dummy_1 - dummy_3 \link{factor} dummy variables
}
}
\keyword{datasets}
