% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdc_descriptives_DT}
\alias{sdc_descriptives_DT}
\title{Example data for \code{sdc_descriptives()}}
\format{
A data.table with 20 rows and 5 columns.
}
\usage{
data("sdc_descriptives_DT")
}
\description{
Utilized in the vignette.
}
\details{
The data.table contains the following columns:
\itemize{
\item id \link{factor} random identifier
\item sector \link{factor} economic sector
\item year \link{integer} time variable
\item val_1, val_2 \link{numeric} value variables
}
}
\keyword{datasets}
