% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_extreme.R
\name{sdc_extreme}
\alias{sdc_extreme}
\title{check if calculation of extreme values comply to RDSC rules.
If so, function returns average min and max values according to RDSC rules.}
\usage{
sdc_extreme(data, id_var, val_var, by = NULL, n_max = 5, n_min = 5)
}
\arguments{
\item{data}{The dataset (anything which can be coerced to data.table) from
which the extreme values can be calculated.}

\item{id_var}{The name of the id variable as a character.}

\item{val_var}{Character vector of value variables on which descriptives are
computed.}

\item{by}{Grouping variables. Can be provided as in
\code{\link[data.table:data.table]{data.table::data.table()}}.}
}
\description{
check if calculation of extreme values comply to RDSC rules.
If so, function returns average min and max values according to RDSC rules.
}
