% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R, R/sdc_DT.R
\docType{data}
\name{sdc_extreme_DT}
\alias{sdc_extreme_DT}
\title{Example data
Utilitzed in the vignette}
\format{
A data.table with 20 rows and 6 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 6 columns.
}
\usage{
data("sdc_extreme_DT")

sdc_extreme_DT
}
\description{
Example data
Utilitzed in the vignette

Test Dataset for vignette
}
\details{
The data.table contains the following columns:
\itemize{
\item id \link{factor} random identifier
\item sector \link{factor} economic sector
\item year \link{integer} time variable
\item val_1 - val_3 \link{numeric} value variables
}
}
\keyword{datasets}
