% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_log.R
\name{sdc_log}
\alias{sdc_log}
\title{Create Stata-like log files from R Scripts}
\usage{
sdc_log(r_scripts, log_files, replace = FALSE)
}
\arguments{
\item{r_scripts}{\link{character} vector containing the path(s) of the R script(s)
which should be run with logging.}

\item{log_files}{\link{character} vector containing the path(s) of the text
file(s) where the log(s) should be stored.}

\item{replace}{\link{logical} Indicates whether to replace existing log files.}
}
\value{
Invisible \code{NULL}.
}
\description{
This function creates Stata-like log files from R Scripts. It
can handle several files (in a \link{character} vector) at once.
}
