% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc_model.R
\name{sdc_model}
\alias{sdc_model}
\title{check if your model complies to RDSC rules}
\usage{
sdc_model(data, model, id_var)
}
\arguments{
\item{data}{\link{data.frame} The dataset from which the model is estimated.}

\item{model}{The estimated model object. Can be a model type like lm, glm and
various others (anything which can be handled by \code{\link[broom:reexports]{broom::augment()}}).}

\item{id_var}{\link{character} The name of the id variable as a character.}
}
\description{
check if your model complies to RDSC rules
}
