% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdc_model_DT}
\alias{sdc_model_DT}
\title{Example data for \code{sdc_model()}}
\format{
A data.table with 80 rows and 9 columns.
}
\usage{
data("sdc_model_DT")
}
\description{
Utilized in the vignette
}
\details{
The data.table contains the following columns:
\itemize{
\item id \link{factor} random identifier
\item y - x_4 \link{numeric} value variables
\item dummy_1 - dummy_3 \link{factor} dummy variables
}
}
\keyword{datasets}
